+++ Models/AppConfig.cs
@@ -7,6 +7,7 @@ public sealed class AppConfig
     public bool ShowCopyJsonButton { get; set; } = false;
     public bool UseAdvancedJsonPieceMapping { get; set; } = false;
     public List<string> AdvancedJsonPieceOrder { get; set; } = [];
+    public string AdvancedSaveMode { get; set; } = "LooseFiles";
     public bool UseDefaultOutputDir { get; set; } = false;
     public string OutputDir { get; set; } = "";
     public bool UseDefaultListName { get; set; } = false;